refactor(remote-control): Migrate UI to LitElement components and integrate @nanostores/lit
