+++ remote-control/package.json
@@ -21,7 +21,8 @@
     "@capawesome/capacitor-google-sign-in": "^0.1.2",
     "@ionic/core": "^8.6.3",
     "@mediagrid/capacitor-native-audio": "^2.3.1",
-    "lit-html": "^3.3.2",
+    "@nanostores/lit": "^0.2.3",
+    "lit": "^3.3.2",
     "nanostores": "^1.2.0"
   },
   "devDependencies": {
