+++ remote-control/src/js/app.js
@@ -1,19 +1,5 @@
 /*
 This file is part of the radio-pad project.
-https://github.com/briceburg/radio-pad
-
-Copyright (c) 2025 Brice Burgess <https://github.com/briceburg>
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-GNU General Public License for more details.
-You should have received a copy of the GNU General Public License
-along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
 import "@ionic/core/css/ionic.bundle.css";
@@ -29,36 +15,55 @@ import { RadioPadAuth } from "./services/auth.js";
 import { RadioListen } from "./services/radio-listen.js";
 import { RadioControl } from "./services/radio-control.js";
 import { RadioPadPreferences } from "./services/preferences.js";
-import {
-  authStore,
-  controlStore,
-  listenStore,
-  preferencesStore,
-  settingsUiStore,
-  toastStore,
-} from "./store.js";
-import { RadioPadUI } from "./ui/index.js";
+
+// Import our Lit Web Components to register them
+import "./ui/radio-auth.js";
+import "./ui/radio-player-tab.js";
+import "./ui/radio-settings.js";
+
+import { toastStore } from "./store.js";
+import { formatErrorMessage, RegistryRequestError } from "./utils/errors.js";
 
 addIcons(appIcons);
 defineCustomElements(window);
 
-function bindStore(store, render) {
-  render(store.get());
-  return store.subscribe(render);
+async function presentToast(notification) {
+  const toast = document.querySelector("#global-toast");
+  if (!toast || !notification?.summary) return;
+
+  const TOAST_SEVERITY = {
+    danger: { color: "danger", duration: 0, position: "top" },
+    warning: { color: "warning", duration: 0, position: "top" },
+    success: { color: "success", duration: 3000, position: "bottom" },
+  };
+
+  const config =
+    TOAST_SEVERITY[notification.severity] || TOAST_SEVERITY.warning;
+  const detailText =
+    notification.format === "registry"
+      ? RegistryRequestError.format(notification.error)
+      : formatErrorMessage(notification.error);
+  const message = detailText
+    ? `${notification.summary} ${detailText}`.trim()
+    : notification.summary;
+
+  toast.message = message;
+  toast.duration = notification.persistent ? 0 : config.duration;
+  toast.color = config.color;
+  toast.buttons = notification.persistent
+    ? [{ text: "Dismiss", role: "cancel" }]
+    : [];
+  toast.position = config.position;
+  await toast.present();
 }
 
 async function bootstrap() {
-  const copyTokenAvailable = !Capacitor.isNativePlatform();
   const prefs = new RadioPadPreferences();
   const auth = new RadioPadAuth();
   const listen = new RadioListen();
   const control = new RadioControl();
-  const ui = new RadioPadUI({ copyTokenAvailable });
 
-  const controlActions = createControlActions({
-    control,
-    listen,
-  });
+  const controlActions = createControlActions({ control, listen });
   const settingsActions = createSettingsActions({
     prefs,
     auth,
@@ -71,28 +76,34 @@ async function bootstrap() {
       settingsActions.refreshAccountsForCurrentRegistry,
   });
 
-  ui.init({
-    onSettingsEdited: () => settingsActions.markDirty(),
-    onSaveSettings: (settingsMap) => settingsActions.save(settingsMap),
-    onSignIn: () => authActions.signIn(),
-    onSignOut: () => authActions.signOut(),
-    onCopyToken: () => authActions.copyToken(),
-    onClickStation: (...args) => controlActions.clickStation(...args),
-    onStopStation: (tabName) => controlActions.stopStation(tabName),
+  // Attach global event listeners from the Lit web components
+  document.addEventListener("auth-signin", () => authActions.signIn());
+  document.addEventListener("auth-signout", () => authActions.signOut());
+  document.addEventListener("auth-copytoken", () => authActions.copyToken());
+
+  document.addEventListener("station-click", (e) =>
+    controlActions.clickStation(e.detail.tabName, e.detail.stationName),
+  );
+  document.addEventListener("station-stop", (e) =>
+    controlActions.stopStation(e.detail.tabName),
+  );
+
+  document.addEventListener("settings-edited", () =>
+    settingsActions.markDirty(),
+  );
+  document.addEventListener("settings-save", (e) =>
+    settingsActions.save(e.detail),
+  );
+
+  // Subscribe to toast store natively here since it requires invoking methods on Ionic's `<ion-toast>` element
+  let lastToastId = 0;
+  toastStore.subscribe((notification) => {
+    if (notification.id !== lastToastId) {
+      lastToastId = notification.id;
+      presentToast(notification);
+    }
   });
 
-  for (const [store, render] of [
-    [preferencesStore, ({ definitions }) => ui.renderPreferences(definitions)],
-    [authStore, (state) => ui.updateAuthState(state)],
-    [settingsUiStore, ({ saveState }) => ui.setSettingsSaveState(saveState)],
-    [controlStore, (state) => ui.renderTabState("control", state)],
-    [listenStore, (state) => ui.renderTabState("listen", state)],
-    [toastStore, (notification) => ui.presentNotification(notification)],
-  ]) {
-    bindStore(store, render);
-  }
-
-  // Single initializer handles offline store boot, auth state, and network fallback
   const wasOauthCallback = await settingsActions.initialize();
   if (wasOauthCallback) {
     const settingsTab = document.querySelector(
